+++ src/services/StateManager.js
@@ -12,7 +12,7 @@ import Logger from '../Logger.js'
  * @return {Promise}
  */
 export function persist(enabled) {
-	const url = generateUrl('/apps/twofactor_email/personal_settings/state')
+	const url = generateUrl('/apps/twofactor_email/personal/save')
 	const data = {
 		state: enabled,
 	}
@@ -27,5 +27,37 @@ export function persist(enabled) {
 			}
 		}).catch(_ => {
 			return { error: 'save-failed' }
+		})
+}
+
+export function persistAdminSettings(settings) {
+	const url = generateUrl('/apps/twofactor_email/admin/save')
+
+	Logger.debug('sending two-factor email admin settings', settings)
+	return Axios.post(url, settings)
+		.then(resp => {
+			if (resp.status !== 200) {
+				return { error: 'save-failed' }
+			} else {
+				return resp.data
+			}
+		}).catch(_ => {
+			return { error: 'save-failed' }
+		})
+}
+
+export function resetAdminSettings() {
+	const url = generateUrl('/apps/twofactor_email/admin/reset')
+
+	Logger.debug('resetting two-factor email admin settings to defaults')
+	return Axios.post(url)
+		.then(resp => {
+			if (resp.status !== 200) {
+				return { error: 'reset-failed' }
+			} else {
+				return resp.data
+			}
+		}).catch(_ => {
+			return { error: 'reset-failed' }
 		})
 }