+++ vite.config.js
@@ -7,6 +7,7 @@ import { createAppConfig } from '@nextcloud/vite-config'
 
 export default createAppConfig({
 	personal_settings: 'src/PersonalSettings.js',
+	admin_settings: 'src/AdminSettings.js',
 	login_setup: 'src/LoginSetup.js',
 	login_challenge: 'src/LoginChallenge.js',
 }, {